backend/ai_model/requirements.txt
joblib==1.4.2
numpy==2.1.2
scikit-learn==1.6.1
joblib==1.4.2
pandas==2.2.3
xgboost==3.0.2
python-dotenv==1.0.1
